backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==3.17.4
python-docx==1.1.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.26.3
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
